lua503 allpass lua515 allpass lua524 allpass lua536 allpass lua546 failed 17 Total failed 17 of 1334 tests, skipped 0 tests.
Modifications to the TForBlock.cs may be wrong

+++ Parser/LHeaderType.cs
@@ -244,7 +244,7 @@ namespace LuaDec.Parser
         }
 
         private static readonly byte[] luacTail = {
-            0x19, (byte) 0x93, 0x0D, 0x0A, 0x1A, 0x0A,
+            0x19, 0x93, 0x0D, 0x0A, 0x1A, 0x0A,
         };
 
         protected static readonly double TEST_FLOAT = 370.5;
